refactor(Participant): 중복 메서드 분리
