+++ src/main/java/domain/Player.java
@@ -11,12 +11,4 @@ public class Player extends Participant {
     public Player(final PlayerDto playerDto) {
         this(new Name(playerDto.name()));
     }
-
-    public boolean isBust() {
-        return hand.sum() > 21;
-    }
-
-    public boolean isBlackjack() {
-        return hand.sum() == 21;
-    }
 }
